requirements.txt
# requirements.txt

# Core web framework for building the user interface
streamlit==1.37.0

# Database connector for MySQL 8.0+
mysql-connector-python==8.4.0

# For secure password hashing
bcrypt==4.1.3

# For processing PDF files (imported as 'fitz')
PyMuPDF==1.24.7

# For processing .docx files (Word documents)
python-docx==1.1.2

# For creating charts and visualizations
matplotlib==3.9.1

# For NLP: keyword extraction, lemmatization
spacy==3.7.5

# Core LangChain libraries for AI integration
# Updated to latest stable, available versions
langchain==0.2.5
langchain-core==0.2.9

# Google AI specific library for LangChain
langchain-google-genai==1.0.7